requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
python-binance==1.0.19
requests==2.31.0
eventlet==0.33.3
python-engineio==4.7.1
python-socketio==5.9.0